Don't delete pinned commands.

+++ crossbow/lib/widgets/command_list_tile.dart
@@ -77,22 +77,27 @@ class CommandListTileState extends ConsumerState<CommandListTile> {
       data: (final valueContext) {
         final projectContext = valueContext.projectContext;
         final command = valueContext.value;
+        final pinnedCommand = valueContext.pinnedCommand;
         return CallbackShortcuts(
           bindings: {
             deleteHotkey: () async {
               if (widget.nullable) {
-                await intlConfirm(
-                  context: context,
-                  message: Intl.message(
-                    'Are you sure you want to delete this command?',
-                  ),
-                  title: confirmDeleteTitle,
-                  yesCallback: () async {
-                    Navigator.of(context).pop();
-                    await projectContext.db.utilsDao.deleteCommand(command);
-                    widget.onChanged(null);
-                  },
-                );
+                if (pinnedCommand != null) {
+                  await intlConfirm(
+                    context: context,
+                    message: Intl.message(
+                      'Are you sure you want to delete this command?',
+                    ),
+                    title: confirmDeleteTitle,
+                    yesCallback: () async {
+                      Navigator.of(context).pop();
+                      await projectContext.db.utilsDao.deleteCommand(command);
+                      widget.onChanged(null);
+                    },
+                  );
+                } else {
+                  widget.onChanged(null);
+                }
               }
             }
           },
